server/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
cryptography>=42.0,<45.0
